+++ packages/babel-plugin-emotion/test/macro/babel-macros-register.js
@@ -2,7 +2,7 @@ require('babel-register')
 const path = require('path')
 require('module-alias').addAliases({
   'emotion-utils': path.join(__dirname, '../../../emotion-utils/src'),
-  'emotion-react/macro': path.join(__dirname, '../../src/macro-styled'),
+  'react-emotion/macro': path.join(__dirname, '../../src/macro-styled'),
   'emotion/macro': path.join(__dirname, '../../src/macro')
 })
