--- packages/emotion-react/src/macro.js
@@ -1,2 +1,0 @@
-// This is only for tests
-module.exports = require('babel-plugin-emotion/src/macro-styled')
