+++ packages/emotion-server/test/index.test.js
@@ -3,7 +3,7 @@
 */
 import React from 'react'
 import { renderToString } from 'react-dom/server'
-import styled from 'emotion-react'
+import styled from 'react-emotion'
 import {
   css,
   injectGlobal,
