+++ packages/react-emotion/test/index.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import serializer from 'jest-glamor-react'
 import { css, sheet } from 'emotion'
-import styled from 'emotion-react'
+import styled from 'react-emotion'
 import { ThemeProvider, withTheme } from 'theming'
 import { mount } from 'enzyme'
 import enzymeToJson from 'enzyme-to-json'
